Fix incompatibiltiies with pytorch 2.1.1 as torch.nn.interpolate does not work with numpy integers anymore

+++ src/controlnet_aux/zoe/zoedepth/models/base_models/midas.py
@@ -99,15 +99,15 @@ class Resize(object):
         self.__resize_method = resize_method
 
     def constrain_to_multiple_of(self, x, min_val=0, max_val=None):
-        y = (np.round(x / self.__multiple_of) * self.__multiple_of).astype(int)
+        y = int(np.round(x / self.__multiple_of) * self.__multiple_of)
 
         if max_val is not None and y > max_val:
-            y = (np.floor(x / self.__multiple_of)
-                 * self.__multiple_of).astype(int)
+            y = int(np.floor(x / self.__multiple_of)
+                 * self.__multiple_of)
 
         if y < min_val:
-            y = (np.ceil(x / self.__multiple_of)
-                 * self.__multiple_of).astype(int)
+            y = int(np.ceil(x / self.__multiple_of)
+                 * self.__multiple_of)
 
         return y
